[+gpu_requirements_user.txt] to specify cuda
We will have instructions in the readme that will specify how the user
can set up the environment on their own local pc. In order to be able to
work with torch and your gpu, we need to install it in the same pip
statement as all the other requirements. Therefore the user can adjust
this user file to their cuda version, and run the command as presented
in the ReadMe.