[Dockerfile] Combined the dockerfiles to merge repos
This new dockerfile sets up the environment with the new combination of
requirements (see commit d64544ced19fdf1a3b06a2df6ffcb0f02bade3ca
Package dependency fix (issue #195)). It is one evironment that works
for both data-management and object-detection, so that the repos can be
merged.

The dockerfile contains some comments about why certain lines are there.
We first build an image to build the ffmpeg from scratch. Then we
restart with a new image (to save memory) and simply copy the end
result.

One big difference with the previous docker files is that all RUN
commands are now done at once, so that we avoid creating extra layers
all the time, since that increases the image size.