[01_check, kso_utils] move .label outside of function
See commit message  caf70d5537e17f4e153db1f931a38a4dcecdeddc from
kso_utils.

the .label in the widgets.open_csv that is used in the notebook is moved
out of the function and not called in the notebook. Without this,
testing the functions will not be possible.