[auto-tests.py] start of the test file
This is a start of the file that will test all notebooks. It is commited
so that it is available for Jannes to continue working on it.

The idea is that we make 1 file that tests all the notebooks. Therefore,
the beginning of the notebooks does not need to be tested every time,
only the things that are special for the specific notebook need to be
noted down.

We have talked about that it is not possible to automatically generate a
test from the notebook, since that does not work with user input. The
idea now was to make sure that the headings in the notebooks correspond
to headings in the tests, so that it is easy to keep it up to date. That
means that when we add new sections to the notebook, that we will need
to manually update the auto-tests.py as well.