[tutorials] change main to master in git clone google colab
The notebooks contain a git clone for in case we use google colab. there
it should clone the master branch. All notebooks now refered to main,
which was in the old repository structure the name of the main/master
branch. But since it is now master, all clone comments are adjusted.