Feature: Add automated testing for notebooks and widgets (#218)
* [01_check, kso_utils] move .label outside of function

See commit message  caf70d5537e17f4e153db1f931a38a4dcecdeddc from
kso_utils.

the .label in the widgets.open_csv that is used in the notebook is moved
out of the function and not called in the notebook. Without this,
testing the functions will not be possible.

* [auto-tests.py] start of the test file

This is a start of the file that will test all notebooks. It is commited
so that it is available for Jannes to continue working on it.

The idea is that we make 1 file that tests all the notebooks. Therefore,
the beginning of the notebooks does not need to be tested every time,
only the things that are special for the specific notebook need to be
noted down.

We have talked about that it is not possible to automatically generate a
test from the notebook, since that does not work with user input. The
idea now was to make sure that the headings in the notebooks correspond
to headings in the tests, so that it is easy to keep it up to date. That
means that when we add new sections to the notebook, that we will need
to manually update the auto-tests.py as well.

* Implement changes to allow testing

* Add step to run notebook tests during linting
* Create test folder structure with notebook-tests.py and widget-tests.py
* Add functions to test for tutorials 1-5. Parts that are commented out
  are currently untested.
* Many tests involving async functions now use a test keyword argument
  to separate testing. This should be improved in future so that the
  same code is tested for both scenarios.
* Pass GH secrets to testing job
* Add GHCR deployment workflow
* Add requirements from kso_utils
* Describe each step within each test workflow
* Note limitations of test workflows (not currently covering
  interactivity, t9 or t10)

---------

Co-authored-by: Diewertje11 <diewertje@kpnplanet.nl>