Dockerfile: HACK: add jupyter to dependencies.
We don't know why do we enable the extensions. But they fail due to
missing modules. It turns out something obscurely changed between
master and dev, and we tracked it down to a change in the
"jupyter==1.0.0" dependency in kso_utils. Until the reason for that
has been figured out, let's just add it manually.