[-build-container.yml,+build-and-test-container.yml]: also run tests
Initially the idea was to create 3 workflows: build-container,
run-tests, update-container. But due to limitations in the github
workflows, everything is merged together in 1 workflow. The limitations
were:
* The test get triggered via the 1st one (workflow_run) and by a change
  in something in the code, causing it to run double. There is no actual
  path-ignore for this.
* Triggering the update-container via 2 workflow runs via part 1 and 2,
  causes that we cannot have the branch or target branch information
  anymore. This makes it hard to update the correct container.

The current approach has 1 workflow that builds the container and runs the tests. It is run all the time, but in PRs we do
not build the container unnecessarily.

This is not the most optimized way to execute things. Since we first, build,
push, and then run the tests. Meaning that if the tests execute, the container
was already pushed. But it is nonetheless the most straight-forward way
to execute the whole process. Dealing with the container is very, very
annoying due to its size and complexity, and github actions are extremely
rigid and have many limitation with respect of when to run the workflows,
and dealing with containers in general. So as a first approach on which to
further iterate, this should certainly work.

Codeveloped-by: Pablo Correa Gomez <pablo.correa.gomez@combine.se>