[build-and-test-container] always push new master
The docker image contains a copy of the code that is used to run things
on SNIC for the users that do not do development. Since we want this to
be up to date with the last commits in master, we always build a new
image when we push to master, even if nothing was changed to the
requirements or the docker file.

It can be considered for the future to work more with releases, so that
we do not need to push new images. See issue #234.