[code-quality.yml] add pylint test for errors
We introduce a part of the tests that pylint can do, so that we check on
whether all functions have the correct amount of arguments, if all
imports are used, if the names of functions actually exist, etc.

These tests will make reorganising code a lot easier.

Later the functionalities of pylint can be expanded by including more of
its test so that we can improve the code quality. But for now we only
keep it to the essentials.