fix: update first two cells of notebooks
I added a roadblock to restart the colab session if panoptes can't be imported, this avoids issues with the requests packages.
I also got rid of the code to fix libmagic issues as it seems no longer needed, this was initially added to deal with panoptes issues.
I combined the installation requirements into the first notebook cell and left the second cell to only load packages.
I added a small description before the first cell for colab users