Remove the double kso_utils directory and adjust CI and pylint (#280)
* [kso_utils] remove double kso_utils in kso_utils

Instead of having everything from the submodule, we now only keep the
actual code that is in the kso_utils.

* [tutorials] change path kso_utils in all nb

Since we now do not have a folder kso_utils wihtin the folder kso_utils
anymore, we can delete this path reference and simplify it. Also the
installing of the requirements is changed in path, since they now live
in the main kso folder.

* remove double kso_utils path in CI and tests

* [build-and-test-container] add pylint

Add the pylint calls to the build-and-test-container, since we want to
run them in an updated container. And since the NB-tests already open
this container, it is added as a step in this job so that we do not need
to do any steps double.

The pylint tests are added as separate steps for the kso_utils, the
tests and the tutorials, since we need a different package (nbqa) to do
pylint on the jupyter notebooks. These 3 will always run (due to the if
success/failure statement). However, the actual notebook and widgets
tests will only run when the pylint tests have passed. They do not need
to run if we have detected errors with pylint.

* fix syntax