[tutorials] add print statement
This print statement is added so that it is possible to correctly
convert the notebooks into .py files with the help of jupytext. The if
statement would be empty without this print statement, since it does not
recognise the other syntax of git clone and pip install. And an empty if
statement is invalid code.

Using jupytext to transfer the the content to a .py file is needed to
use vulture to search for dead code, since vulture cannot look inside
the notebooks. And since the code mostly gets used within the notebooks,
it is needed to have them in .py format.

These .py files will not be added to the repository, but simply
teporarily created on a gitlab runner by the CI.