[+detect-unused-code, pyproject.toml] CI detect unused code
This commits implements a new part of the CI where we use Vulture to
detect unused code. This check is allowed to fail and will always show
as pass. However, we can go in into the check and see the output with
which functions/variables/attributes/etc are likely to be unused.