[+whitelist, detect-unused-code.yml]
The whitelist.py is a file in which we can state all errors Vulture
reports that we have checked and want to keep. Vulture states many false
positives and therefore this is needed.

With this file we can also have the vulture CI check fail when it
actually detects unused code.

So the idea is to check everything and put all false positives in the
whitelist. Once in the so much time we should clear the whitelist and
recheck everything.