Yolov5 (#296)
* [README] update to reflect yolov5 tag of repo

Now that we are migrating to Cloudina, we need to also migrate to Yolov8
and it is not possible to keep both around due to package dependencies.
Therefore we now create a yolov5 tag in the repo for if someone wants to
keep using yolov5.

The repo with yolov5 will therefore not get further develloped. All
development will be only available for yolov8.

This commit changes the README to reflect this decision and changes the
instructions.

* [Dockerfile] correct path requirements

Apparently we had forgotten to correct the path tho the requirements
file when we removed the submodule. This is done now.