fix: aggregate_detections and gbif export
I updated the aggregate_detections func so that now it works with project-specific information (e.g. sql movie and site information) and formats the detections df for multiple uses (e.g. not only calculate maxN)

In project:
choose_footage now also stores a map of the moives names and movies ids selected
added the t9-specific function to download ml aggregated detections
commented out the runs limit as it messed up with the model selection (maybe it was a short fix?)

In tutorial_utils:
updated the export to gbif_occurence

In tut#8 #9:
updated the aggregate labels and export to gbif arguments