fix: changed the logic of clip extractions
Attempt to streamline and simplify the way we extract clips as there was a confusing mix of example vs non example clips, modified vs non modifed clips and ffmpeg vs python_ffmpeg.

The new logic uses the "is_example" argument of the function "create_clips" to generate example or "ready to upload to Zooniverse" clips.

In tutorial_utils and movie_utils:
I used ffmpeg_python instead of subprocess to create clips and deleted the "create_modified_clips" and "review_clip_selection" functions as they're no longer used.

I also set the clips to remove audio as default.

For easier formatting, I moved the clip-related functions close to each other.

In widgets:
I updated the way we display the number of clips to be generated