[MLProjectProcessor,general]remove dynamic imports
Since we have now split out the MLProjectProcessor to its own file, and
all these imports are performed during the initialisation of the class,
it is better to just have them at the top.

We also remove the function import_modules in general.py, since this is
now no longer used.