REMOVE: mlflow
Right now neither wandb and mlflow is working completely. And all the
logic is spread out over the entire code base. We would like this to be
more modular and take this out and have for now decided to only do that
for wandb. And remove the usage of mlflow for now.