[MLProjectProcessor, general] set registry=wandb and remove modules
The registry that is used to store models and do the tracking of the
training process was not set at all before, that is why we needed to add
it manually to the notebook when we wanted to run things. Now it is set.
However, it should not be set here, it should become a config in the
future config file, and here it should be read from there. But till we
have it, it is hardcoded here.

The commit also removes the importing of the model modules, this was not
used anywhere later on and we do not want to have these hardcoded
imports left in the code. This peice of code also never got executed in
the state the code was in now, since registry was None.