REFACTOR: move wandb code to separate wandb_utils
We want to make the code more modular. The idea is that every registry
will have its own utils file with this same set of functions, so that
they can be used depending on which registry it is.

Right now there are still if statements left, but the plan is to call
upon these functions with getattr later.