[MLProjectProcessor, general, wandb_utils] module
Use the wandb_utils as a module that can be imported, so that we do not
need to mention the registry at all in the MLProjectProcessor itself.
(there are 2 instances left, but I do not understand the code there yet,
so this needs to be looked at later)

Also add a validation functions for this utils file, so that when we
would create one for another registry, that we make sure that it
contains those functions we need.

This commit also adds correct opening and closing of the runs in WandB
where it does not happen automatically via YOLO.