Refactor: move wandb_utils code to the registries folder
To make it clear where all code related to different registries modules
should be, like WandB or MLflow, we now have a separate folder for
these.

We keep the name wandb_utils.py and do not call it wandb.py, since
python would in that case get confused when importing and uses this file
instead of the wandb package.