REFACTOR: [-general.py, +test_registries.py]
The funcitons in the file general.py were only used in 1 place, so they
are moved there instead.
Also, the check of if a registry contains all necessary functions should
not be done in the code we run, but should be checked by CI. So the file
test_registries is created. No test is written there yet, but that will
come in a next commit.