[+test_registries, wandb_utils] simplify modelfunc
This commit changes the interaction with the models from the registries.
We do not want to have any widget functionality here, because the model
name should come from the config file. So we only need a get_model
function.
To keep a bit of the user friendliness from the widgets, we created a
show_available_models function that returns a list with all the models.
My idea here is that the notebook can contain one cell that runs
everything needed to train, so calls upon a python script with all that
code. That uses a config file as input.
Other cells below can contain functions like the show_available_models,
to help the user decide on what to put in the config file. But they are
not needed or used in the training itself.
So that this training code could also be called upon via a bash script
for example.

This code also adds tests for these get_model and show_available_models
functions.
It also has a first sketch for a test for the get_dataset, but I don't
know any example of a model where this code would work, so this one does
not really test anything yet and needs to be added to later.

Note: this commit is in slightly incomplete, since it does not remove
the usage of choose_model, choose_baseline_model and get_model in the
MLProjectProcessor and the notebooks. This will be done in the coming
commits. To split this up is a concious choice to keep this commit small
and not change entire notebooks too.