[+train_models.py, test_train_models, train_models.ipynb]
This commit shows how I envision the future notebooks to look like. The
python script train_models can be called from the notebook or from a
bash script or from other places. The notebook has some description and
an option to set the parameters with which the train_models.py gets
called.

The test currently only looks if it runs and it does!
However, it does not log the result on WandB yet, only in a folder
locally, so this needs to be investigated.