correct usage of get_model, choose_baseline_model, choose_model
In a previous commit, I removed the choose_... functions, since we do
not want this interactivity here but want it to come from a config. That
commit did not change the usage in the notebooks yet. That is corrected
here.

One function had zenodo functionalities. This is now covered in a
separate get_and_show_models_zenodo. This functionality we want for user
friendliness, but this will need to be moved and renamed later to get
modularity and be able to swap zenodo out for other services. That is
work for later.