Remove widgets, instead take input by code
This commit removes many widgets, since we do not want this
interactivity. The widgets are hard to maintain and test. Instead, the
user can set these variables themselves by text in the notebook, or in a
config file if we want to run the code via for example a bash script.
The notebooks are adjusted to this structure.

The widgets that are removed in this commit:
choose_entity, choose_train_params, choose_project,
choose_experiment_name, choose_conf, choose_text, choose_test_prop,
choose_eval_params, choose_folder (usage in notebooks is removed, the
function still exists as _choose_folder, since it is used inside another
widget)