[MLProjectProcessor, wandb_utils, github_ci] login to wandb
If you have logged in once on your local computer, you do not need to
log in again. However, CI still needs to log in when it runs, using the
gitlab secret WANDB_API_KEY.
When running on a server, this probably also needs to be used.

To be able to use the secret in CI, add this to the yaml file.