[requirements.txt, Dockerfile, ...] update
Many of the requirements were very old versions of different packages.
This commit updates everything to the newest version and removes
packages that are not needed anymore.
All tests that run in CI still work, so those core parts of SUBSIM work
with this new dependency list. The other code is not checked.

Also, add scikit-learn back in, since this is used in frame_utils. Even
though this is not covered by tests yet, and it is unknown if this will
be part of the core of SUBSIM, not having this dependency makes pylint
fail. So add it back in for now.

This update from python 3.8 to 3.12 and the updates of the packages are
necessary to be able to keep up with the changes and to keep it
maintainable.