[Dockerfile] use conda-forge
Configure conda-forge as the only channel
Since mid 2024, Anaconda, Inc. introduced ToS acceptance for the
default Anaconda channels, meaning that conda create requires
explicit acceptance. Therefore use conda-forge.

We first set conda-forge in the settings, so that if users would install
any packages, they also use conda-forge. However, to make it work in the
dockerfile, we also need to add --override-channel to every conda call.

Note:
We use some packages that cannot be resolved by conda, therefore we need
pip. Unless we want to devide the requirementsfile into 2, one for pip
and one for conda, it is not possible to first try to install with conda
and then with pip. So instead we keep using this solution we already
had: we install everything with pip, and only opencv, where we know that
there are some system dependencies we need, we install it with anaconda.