[Dockerfile] remove SNIC specific configuration
This kind of HPC specific configuration should be set in the
configuration that is used by the HPC cluster when they start the
containers in their system. This configuration will be different on SNIC
and LUMI and should therefore not live in the container itself.

The DATA_DIR and ARTIFACT_DIR are not used anywhere in the code, so try
if it works without.