[Dockerfile] use pip, no conda, update base image
Using conda has some advantages:
- we do not need to define the system libraries ourselves in the docker
  file, these are included in the conda install op the python package
- GPU/ML packages often have CUDA-aware conda builds, while with pip we
  would need to specify it manually.
- Researchers might freely add packages to the requirements.txt file,
  which might need extra system libraries and therefore would not work
  if using pip. With conda, these system libraries would automatically
  be installed as well.

I thought that this last reason is most relevant for us, however, since
we first install everything with pip and need to separately define what
we will install with conda, we would need to change this manually here
anyway. The only difference is that when you use pip, we need to look up
ourselves which system libraries are needed.

But since pip is more reliable in a production setup, since for conda
things underneath the hood might change without notising, this commit
does an attempt to see if everything works with just the use of pip.

update base image to newer version

The ubuntu image we were using is only supported till 2025. And it does
not contain python 3.12, so we cannot keep using it.
When we used conda, this was not a problem, since conda automatically
installed all system dependencies that were needed for python 3.12, but
when using pip, this is not the case and we need the correct base image.

Also, the 2 variables removed here were conda specific and should be
removed when moving to pip.

Explanation for the choice of the image:
SNIC has updated their cuda to 13.0, meaning that all cuda versions
<13.0 can be used for the container.
Pytorch has as latest stable version 2.8.0, which can work with cuda
12.6, 12.8 and 12.9. The newer the better they describe.
Pytorch 12.9 works with cdnn 9.

Also, we want Ubuntu24.04 so that we have python 3.12.
Therefore we get this as the new base image.