[requirements,build-and-test-container]CI packages in container
We are running into permission errors when trying to install packages in
the CI step itself, this is because we set a user at the end of the
docker container, causing that you do not run as root. We do not want to
run as root for safety reasons, but this causes that we cannot install
packages. Even not with 'pip install --user'.

Users should be able to install extra packages with 'pip install --user'
from the notebooks. However, this we should test to make sure that that
works.